Fixed bug
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@12327 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ daq/slc/system/src/Buffer.cc
@@ -7,7 +7,7 @@ using namespace Belle2;
 Buffer::Buffer()
 throw() : m_memory(NULL), m_size(0), m_allocated(false) {}
 
-Buffer::Buffer(size_t size, unsigned char* memory) throw()
+Buffer::Buffer(unsigned int size, unsigned char* memory) throw()
   : m_memory(memory), m_size(size), m_allocated(false)
 {
   if (memory == NULL) {
